requirements.txt
PyYAML==3.12
logging==0.4.9.6
coincheck==0.1.10
pytz==2016.7
requests==2.6.0
simplejson==3.8.1
